Add UniformTypeIdentifier.localizedDescription (#5)
Previously, `UniformTypeIdentifier.description` was calling `UTTypeCopyDescription()`, which is a very heavy function that seems to make interprocess method calls. That was too expensive for a relatively common call like `description`. So, that functionality has been moved to `UniformTypeIdentifier.localizedDescription`, and `description` now just returns the raw value of the type.